Fix: Ad size in iOS setting wrong height #112
This commit fixes a bug in the `toCommon()` extension function for `GADAdSize` on the iOS platform.

Previously, when converting a `GADAdSize` to the common `AdSize` class, the `height` property was incorrectly being assigned the value of the ad's `width`.

This has been corrected to use the `height` from the `GADAdSize`, ensuring that the dimensions of the ad size are mapped correctly.